NexoraAI/monolith_backup/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pandas
scikit-learn
joblib
httpx
beautifulsoup4
python-jose[cryptography]
passlib[argon2]
argon2-cffi
pydantic[email]
email-validator
groq
redis==7.2.0
xgboost
